requirements.txt
dash==3.0.4
ipykernel==6.29.5
nbformat==5.10.4
pandas==2.2.3
pycoingecko==3.2.0
pymongo==4.13.0
python-dotenv==1.1.0
scikit-learn==1.7.0
APScheduler==3.11.0